gcc: Package libstdc++ gdb python helpers into dev package
People are seeing these errrors from ldconfig
libstdc++.so.6.0.14-gdb.py is not an ELF file - it has the wrong magic
bytes at the start.

this file is moved into gdb's autoload directory
if it exists there then gdb will find it when debugging
and it wont be in the paths where ldconfig looks for
libraries.

gcc-package-cross.inc:  Don't package libgcc-dev since it conflicts with libgcc-s-dev

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Tested-by: Philip Balister <philip@balister.org>
Signed-off-by: Steve Sakoman <steve@sakoman.com> for gcc-package-cross.inc
Signed-off-by: James Ang <angweiyang@gmail.com>